sg_fill_brewer {streamgraph}	R Documentation

Use ColorBrewer palettes for streamgraph fill colors

Description

ColorBrewer provides sequential, diverging and qualitative colour schemes which are particularly suited and tested to display categorical values.

Usage

sg_fill_brewer(sg, palette = "Spectral")

Arguments

sg
streamgraph object

palette
ColorBrewer pallete atomic character value (defaults to Spectral)

Value

streamgraph object

Examples

## Not run: 
library(dplyr)
library(streamgraph)
ggplot2movies::movies %>%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) %>%
  tidyr::gather(genre, value, -year) %>%
  group_by(year, genre) %>%
  tally(wt=value) %>%
  ungroup %>%
  mutate(year=as.Date(sprintf("%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") %>%
  sg_fill_brewer("PuOr")

## End(Not run)

[Package streamgraph version 0.9.0 Index]