sg_fill_manual {streamgraph}	R Documentation

Use manual colors for streamgraph fill colors

Description

Specify a vector of colors (e.g. c("red", "#00ff00", rgb(0,0,1))) to use for the color scale. Note that streamgraph sorts the categorical values before assigning the mappings, which means you can use that as a determinstic way of assigning specific colors to categories. If the number of categories exceeds the number of colors in the palette, the colors will be reused in order.

Usage

sg_fill_manual(sg, values = NULL)

Arguments

sg
streamgraph object

values
character vector of

Value

streamgraph object

Examples

## Not run: 
library(dplyr)
library(streamgraph)
ggplot2movies::movies %>%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) %>%
  tidyr::gather(genre, value, -year) %>%
  group_by(year, genre) %>%
  tally(wt=value) %>%
  ungroup %>%
  mutate(year=as.Date(sprintf("%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") %>%
  sg_fill_manual(c("black", "#ffa500", "blue", "white", "#00ff00", "red"))

## End(Not run)

[Package streamgraph version 0.9.0 Index]